pyproject.toml
[project]
name = "Exo_Skryer"
description = "A JAX accelerated retrieval code"
readme = "README.rst"
requires-python = ">=3.12"
license = { file = "LICENSE", content-type = "text/plain" }
authors = [
  { name = "Elspeth K. H. Lee", email = "elspeth.lee@unibe.ch" },
]
dependencies = [
  "numpy",
  "jax",
  "blackjax @ git+https://github.com/handley-lab/blackjax.git@95e93ed93e726b95143b2f4a9582888621625f52",
  "jaxns",
  "numpyro",
  "arviz",
  "pyyaml",
  "optimistix",
  "anesthetic",
  "distrax",

  # below is the recommended source for tensorflow_probability:
  "tfp-nightly"
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest-doctestplus",
    "pytest >=4.6.0",
    "tox"
]
docs = [
    "sphinx",
    "sphinx_astropy",
    "sphinx-automodapi",
    "tomli; python_version <\"3.11\"",
    "sphinx-astropy[confv2]",
    "sphinx-rtd-theme",
    "sphinx-toolbox",
    "matplotlib",
    "tox"
]

[project.urls]
repository = "https://github.com/eleeastro/exo_skyer"

[build-system]
requires = [
  "setuptools>=61.2",
  "setuptools_scm[toml]>=3.4",
  "wheel",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]

[tool.setuptools_scm]
write_to = "exo_skryer/version.py"

[tool.pytest.ini_options]
testpaths = [
  "exo_skyer",
  "docs",
]
doctest_plus = "enabled"
text_file_format = "rst"
addopts = "--doctest-rst"

[tool.coverage.run]
omit = [
 "exo_skyer/conftest.py",
 "exo_skyer/*setup_package*",
 "exo_skyer/tests/*",
 "exo_skyer/*/tests/*",
 "exo_skyer/extern/*",
 "exo_skyer/version*",
 "*/exo_skyer/conftest.py",
 "*/exo_skyer/*setup_package*",
 "*/exo_skyer/tests/*",
 "*/exo_skyer/*/tests/*",
 "*/exo_skyer/extern/*",
 "*/exo_skyer/version*",
]

[tool.coverage.report]
exclude_lines = [
  # Have to re-enable the standard pragma
  "pragma: no cover",
  # Don't complain about packages we have installed
  "except ImportError",
  # Don't complain if tests don't hit assertions
  "raise AssertionError",
  "raise NotImplementedError",
  # Don't complain about script hooks
  "def main(.*):",
  # Ignore branches that don't pertain to this version of Python
  "pragma: py{ignore_python_version}",
  # Don't complain about IPython completion helper
  "def _ipython_key_completions_",
]